[update] ロジックをコンポーネントに集約
